Intellisense Progress
Updated _Suggestions partial view to only be the script that gets the
hints for code completion since other scripts won't be changing. Created
a new action that updates that partial view with suggestions for that
Object. Currently have just tested it out with the Console class. Only
have mscorlib assembly in the intellisense but will add the remaining
ones.